Lets-go- Website
SRSC has their own website where we can get all the information about the services provided by them. However, there is no way for the people to connect with each other. If a student is looking for a partner during exercise or any sports at a specific time, there is no way for him/her to find it out. So, we are planning to create a website where students can post which game they want to play and at what time. Other students who are interested in the same game and who are free at that time, can contact their partners. This way the students will be able to search for partners who have similar interests. We are going to call this website “Let's meetup”.